:poop: 企图实现图片预处理加进度条百分显示
