+++ jiuguan/src/router/index.js
@@ -5,7 +5,7 @@ import Login from '../views/Login.vue'
 import Start from '../views/Start.vue'
 import Register from '../views/Register.vue'
 import Community from '../views/Community.vue'
-import LoginWelcome from '../views/LoginWelcome.vue'
+import Loading from '../views/Loading.vue'
 Vue.use(VueRouter)
 
 const routes = [
@@ -30,9 +30,9 @@ const routes = [
     component: Community
   },
   {
-    path: '/loginWelcome',
-    name: 'LoginWelcome',
-    component: LoginWelcome
+    path: '/loading',
+    name: 'Loading',
+    component: Loading
   }
 ]
