--- jiuguan/src/views/LoginWelcome.vue
@@ -1,13 +1,0 @@
-<template>
-  <p>登录成功，欢迎你回到酒馆</p>
-</template>
-
-<script>
-export default {
-
-}
-</script>
-
-<style>
-
-</style>
